@@ -6,7 +6,7 @@
 
     <groupId>it.ozimov</groupId>
     <artifactId>scheduling-and-persistence-example</artifactId>
-    <version>0.6.3</version>
+    <version>3.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>Spring Boot Email - Scheduling and Persistence Example</name>
@@ -18,7 +18,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.5.6.RELEASE</version>
+        <version>3.0.0</version>
     </parent>
 
     <dependencies>
@@ -29,12 +29,12 @@
         <dependency>
             <groupId>it.ozimov</groupId>
             <artifactId>spring-boot-email-core</artifactId>
-            <version>0.6.3</version>
+            <version>3.0.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>it.ozimov</groupId>
             <artifactId>spring-boot-freemarker-email</artifactId>
-            <version>0.6.3</version>
+            <version>3.0.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
 
@@ -44,8 +44,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>11</source>
+                    <target>11</target>
                 </configuration>
             </plugin>
         </plugins>
